back-end/requirements.txt
django==5.0
djangorestframework
environs
markdown
django-filter
